add additional users to head node
In order for a user to submit jobs, it needs to exist, and have the same uid,
on all nodes in the cluster (including the head node). Add the additional
users to the head node.

This change also mounts the user's nfs share on the head node. Applying this
change requires users to re-run the slurm.yml playbook on all head nodes. This
can be done by running `terraform apply`.